+++ src/store_System/ISystem.java
@@ -14,7 +14,7 @@ public interface ISystem {
     public List<Product> searchProductsByName(String name);
     public List<Product> searchProductsByCategory(String category);
     public List<Product> searchProductsByKeyword(String keyword);
-    public List<Product> filterByPrice(List<Product> base , int min , int max);
+    public List<Product> filterByPrice(List<Product> base , double min , double max);
     public List<Product> filterByRating(List<Product> base , int min , int max);
     public List<Product> filterByCategory(List<Product> base , String category);
     public List<Product> filterByStoreRating(List<Product> base , int min, int max);
