+++ src/store_System/System.java
@@ -116,7 +116,7 @@ public class System implements ISystem {
         }
     }
 
-    public List<Product> filterByPrice(List<Product> base , int min , int max){
+    public List<Product> filterByPrice(List<Product> base , double min , double max){
         List<Product> toReturn = new LinkedList<>();
         for(Product p : base){
             if(p.getPrice()>= min & p.getPrice()<= max){
